requirements.txt
numpy
pandas
matplotlib
seaborn
pytorch-lightning
wandb
cupy-cuda12x